requirements.txt
flask
torch
transformers
pillow
pyttsx3